+++ src/tx/fcointx.h
@@ -55,7 +55,7 @@ public:
     uint256 ComputeSignatureHash(bool recalculate = false) const {
         if (recalculate || sigHash.IsNull()) {
             CHashWriter ss(SER_GETHASH, 0);
-            ss << VARINT(nVersion) << nTxType << VARINT(nValidHeight) << txUid
+            ss << VARINT(nVersion) << uint8_t(nTxType) << VARINT(nValidHeight) << txUid
                << toUid << VARINT(llFees) << VARINT(fcoins);
 
             sigHash = ss.GetHash();
