+++ src/tx/mulsigtx.h
@@ -98,7 +98,7 @@ public:
     uint256 ComputeSignatureHash(bool recalculate = false) const {
         if (recalculate || sigHash.IsNull()) {
             CHashWriter ss(SER_GETHASH, 0);
-            ss << VARINT(nVersion) << nTxType << VARINT(nValidHeight);
+            ss << VARINT(nVersion) << uint8_t(nTxType) << VARINT(nValidHeight);
             // Do NOT add item.signature.
             for (const auto &item : signaturePairs) {
                 ss << item.regId;
